Added UAC field to the sampleIterator.
Updated the description in the jobConvertor to take in the telNo and
ContactName as string arrays, loop though and display in each pair as a
single contact. Also to display UAC number.

Updated the NSW field period to look for N<stage> in field period table
in anticipation of different field periods for NSW, in the SampleIngest.